+++ common/src/comp/ability.rs
@@ -717,6 +717,7 @@ impl From<&CharacterState> for CharacterAbilityType {
             | CharacterState::Climb(_)
             | CharacterState::Sit
             | CharacterState::Dance
+            | CharacterState::Pet(_)
             | CharacterState::Talk
             | CharacterState::Glide(_)
             | CharacterState::GlideWield(_)