+++ common/src/states/wielding.rs
@@ -9,6 +9,7 @@ use crate::{
         behavior::{CharacterBehavior, JoinData},
         idle,
     },
+    uid::Uid,
 };
 use serde::{Deserialize, Serialize};
 
@@ -101,6 +102,12 @@ impl CharacterBehavior for Data {
     fn dance(&self, data: &JoinData, _: &mut OutputEvents) -> StateUpdate {
         let mut update = StateUpdate::from(data);
         attempt_dance(data, &mut update);
+        update
+    }
+
+    fn pet(&self, data: &JoinData, _: &mut OutputEvents, target_uid: Uid) -> StateUpdate {
+        let mut update = StateUpdate::from(data);
+        attempt_pet(data, &mut update, target_uid);
         update
     }