+++ common/systems/src/stats.rs
@@ -131,6 +131,7 @@ impl<'a> System<'a> for Sys {
                 CharacterState::Idle(_)
                 | CharacterState::Talk
                 | CharacterState::Dance
+                | CharacterState::Pet(_)
                 | CharacterState::Skate(_)
                 | CharacterState::Glide(_)
                 | CharacterState::GlideWield(_)